change Split files and explain why used expanded
