+++ android/kotlin-profile/2024-02-09-18-34-31-789.profile
@@ -158,3 +158,26 @@ OS_TYPE=Windows 10
 PROJECT_PATH=D:\Real Project\expense_master\android
 STATISTICS_VISIT_ALL_PROJECTS_OVERHEAD=17
 BUILD FINISHED
+BUILD_FAILED=true
+BUILD_FINISH_TIME=1708504368983
+BUILD_SRC_EXISTS=false
+CONFIGURATION_API_COUNT=7
+CONFIGURATION_IMPLEMENTATION_COUNT=7
+CPU_NUMBER_OF_CORES=8
+EXECUTED_FROM_IDEA=true
+GRADLE_BUILD_DURATION=72070
+GRADLE_BUILD_NUMBER_IN_CURRENT_DAEMON=1
+GRADLE_DAEMON_HEAP_SIZE=1690272505
+GRADLE_EXECUTION_DURATION=11371
+GRADLE_NUMBER_OF_TASKS=2435
+GRADLE_NUMBER_OF_UNCONFIGURED_TASKS=1868
+GRADLE_VERSION=7.5.0
+KOTLIN_COMPILER_VERSION=1.7.10
+KOTLIN_KTS_USED=false
+KOTLIN_OFFICIAL_CODESTYLE=false
+KOTLIN_PROGRESSIVE_MODE=false
+NUMBER_OF_SUBPROJECTS=8
+OS_TYPE=Windows 10
+PROJECT_PATH=D:\Real Project\expense_master\android
+STATISTICS_VISIT_ALL_PROJECTS_OVERHEAD=41
+BUILD FINISHED
