+++ lib/core/routing/name_router.dart
@@ -1,5 +1,4 @@
 abstract class NameRouter {
   static const String onboardingView = "/onbordingView";
   static const String homeView = "/homeView";
-  static const String homeView2 = "/homeView2";
 }
